@rth 2017-11-29
I like the idea! So this would speed up the `kneighbours` method at the cost of a somewhat slower `fit`. 

Here are [some additional benchmarks](https://github.com/rth/ipynb/blob/master/sklearn/PR_10212/benchmark_nearest_neighbours_brute_PR10212.ipynb) where I look at the speedup for the `fit` and `kneighbours` methods for randomly selected cases in a parameter space defined by `n_samples`, `n_features`, `n_queries` and `n_neighbours`. 

The summary is that, 
 * For dense arrays, this change indeed speeds up `kneighbours` method by 1.5-2. Best case scenario is 2-3 and worst case is 0.95. On the contrary `fit` is speed up by a factor of 0.45 - 0.7.  Overall the speedup for `fit` + `kneighbours` is 1.1 - 1.3 with the best case scenario at 3 and a worst case at 0.95.
  * For sparse arrays, this change would speed up both `fit` and `kneighbours` by somewhere between 0.25 and 0.6 (i.e. make everything much slower) so it's probably better to keep the current solution for sparse arrays.

Note that this uses your simplified `fast_euclidean_neighbors` function and not the actual code in the PR.

Also, this would ignore the `n_job` parameter with `euclidean` distances, which is reasonable I think,  as it's better to let BLAS handle that. So, I think a warning could be printed when user attempts to compute Euclidean distances on dense arrays with `n_jobs > 1` (which would partially address https://github.com/scikit-learn/scikit-learn/issues/8216).