@jnothman 2017-12-11
Test failures suggest you've got logic errors.

In terms of implementation, can't we just add the norms in `effective_metric_params_`?

In what cases would you recommend `algorithm='brute'` for euclidean? If sparse, please benchmark that case too.