@rth 2017-12-14
> In what cases would you recommend algorithm='brute' for euclidean

In my experience, when computing pairwise distances with `n_features` more than a few 100 , when BallTree wouldn't work so well.