@jnothman 2017-12-14
should this pr also integrate that logic into 'auto'?