@lesteve 2021-05-05
This is now longer relevant, probably something similar has been done by @jeremiedbb work on nearest neighbors.